alfaneo-test-main/requirements.txt
# Core
fastapi
uvicorn[standard]
python-dotenv

# Scraping
selenium
beautifulsoup4

# Agente LLM
langchain
langchain-cloudflare
requests